+++ hrswp-documents.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: HRSWP Documents
- * Version: 1.0.0
+ * Version: 1.0.0-rc.2
  * Description: A WSU HRS WordPress plugin to provide document management.
  * Author: Adam Turner, washingtonstateuniversity
  * Author URI: https://hrs.wsu.edu/
@@ -100,6 +100,9 @@ function get_plugin_info( $type = '' ) {
 			break;
 		case 'plugin_file_uri':
 			$output = __FILE__;
+			break;
+		case 'post_type':
+			$output = 'hrswp_documents';
 			break;
 		case 'slug':
 			$output = 'documents';