+++ lib/load.php
@@ -8,8 +8,15 @@
 
 namespace HrswpDocuments\lib\load;
 
+use HrswpDocuments\inc\Documents_Post_Type;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	die( 'Silence is golden.' );
 }
 
+require dirname( __FILE__ ) . '/client-assets.php';
 require dirname( __FILE__ ) . '/data.php';
+
+require dirname( __DIR__ ) . '/inc/class-documents-post-type.php';
+
+Documents_Post_Type\Documents_Post_Type::factory();